Re-extract fields into a standalone file
